Reorganise and simplify question controller
